\name{eecf}
\alias{eecf}

\title{
Empirical Exposure Curve Function
}
\description{
Compute an empirical exposure curve function.

}
\usage{

eecf(x)


}

\arguments{
\item{x}{
numeric vector of the observations for \code{eecf}; for the methods, an object of class \code{"eecf"}.
}


}
\details{
TODO!

}
\value{
A numeric value or a vector.
}
\references{
TODO
}
\author{
Dutang Christophe
}
\seealso{
\code{\link{exposureCurve}}.
}
\examples{
#TODO

}

