\name{etl}
\alias{etl}

\title{
Empirical total loss
}
\description{
Compute the empirical total loss.

}
\usage{

etl(x)


}

\arguments{
\item{x}{
numeric vector of the observations.
}


}
\details{
TODO!

}
\value{
A numeric value or a vector.
}
\references{
TODO
}
\author{
Dutang Christophe
}

\examples{
#TODO

}

