\name{etl}
\alias{etl}

\title{
Empirical total loss
}
\description{
Compute the empirical total loss.

}
\usage{

etl(x, na.rm=FALSE)


}

\arguments{
\item{x}{numeric vector of the observations.}
\item{na.rm}{a logical value indicating whether \code{NA} values should 
be stripped before the computation proceeds.}

}
\details{
Compute the empirical total loss defined as the proportion of full destruction rates,
that is observations that equal 1.

}
\value{
A numeric value or a vector.
}
\references{
TODO
}
\author{
Dutang Christophe
}

\examples{
#TODO

}

