% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{g2a}
\alias{g2a}
\title{Get a parameter known g and b}
\usage{
g2a(g, b)
}
\arguments{
\item{g}{the g parameter}

\item{b}{the b parameter}
}
\value{
a real value
}
\description{
\code{g2a} returns the a parameter known g and b
}
\examples{
g2a(10,2)
}

