\name{mbbefd-package}
\alias{mbbefd-package}
\alias{mbbefd}
\docType{package}
\title{
 MBBEFD distribution and exposure curve
}
\description{
It contains the definition of MBBEFD distribution (distribution function, density, quantile functions)
as well as MBBEFD exposure curve. The vignette shows code snippets to fit the distribution to empirical data.
}
\details{
\tabular{ll}{
Package: \tab mbbefd\cr
Type: \tab Package\cr
Version: \tab 0.6.3\cr
Date: \tab 2015-03-29\cr
License: \tab GPL 2\cr
}
}
\author{
Giorgio Alfredo Spedicato[cre, aut] and Christophe Dutang [ctb]
Maintainer:  Giorgio Spedicato <spedicato_giorgio@yahoo.it>
}
\references{
BERNEGGER, STEFAN. THE SWISS RE EXPOSURE CURVES AND THE MBBEFD DISTRIBUTION CLASS. Astin Bulletin (1997): 99.
}
\keyword{package}
\seealso{
See \code{\link{mbbefd-distr}} for the MBBEFD distribution;
see \code{\link{swissRe}}, \code{\link{exposureCurve}} for exposure curves;
see \code{\link{oidistribution}}, \code{\link{oibeta}}, \code{\link{oiunif}}, \code{\link{oistpareto}} for one-inflated distributions.
}

